Move CatalogTree related components from ConnectionForm to CatalogTree
