+++ airbyte-webapp/src/views/Connection/ConnectionForm/components/SyncCatalogField.module.scss
@@ -8,52 +8,3 @@
   font-size: 14px;
   line-height: 17px;
 }
-
-%headerCell {
-  font-size: 10px;
-  line-height: 13px;
-}
-
-.checkboxCell {
-  @extend %headerCell;
-
-  max-width: 43px;
-  text-align: center;
-}
-
-.catalogHeader,
-%catalogHeader {
-  min-height: 33px;
-  margin-bottom: 0 !important;
-  margin-left: 20px;
-}
-
-.readonlyCatalogHeader {
-  @extend %catalogHeader;
-
-  margin-left: 50px;
-}
-
-.catalogSubheader {
-  @extend %catalogHeader;
-
-  padding-top: 10px;
-  margin-left: 155px;
-}
-
-.readonlyCatalogSubheader {
-  @extend %catalogHeader;
-
-  padding-top: 10px;
-  margin-left: 115px;
-}
-
-.catalogTreeContainer {
-  margin-bottom: 29px;
-  max-height: 600px;
-  overflow-y: auto;
-
-  --webkit-overlay: true;
-
-  width: 100%;
-}
