+++ airbyte-webapp/src/views/Connection/ConnectionForm/components/SyncCatalogField.tsx
@@ -1,198 +1,52 @@
-import classnames from "classnames";
 import { FieldProps } from "formik";
-import React, { useCallback, useMemo, useState } from "react";
+import React, { useCallback } from "react";
 import { FormattedMessage } from "react-intl";
-import styled from "styled-components";
 
-import { Cell, Header } from "components/SimpleTableComponents";
-import { CheckBox } from "components/ui/CheckBox";
-import { LoadingBackdrop } from "components/ui/LoadingBackdrop";
 import { Text } from "components/ui/Text";
-import { InfoTooltip, TooltipLearnMoreLink } from "components/ui/Tooltip";
 
 import { SyncSchemaStream } from "core/domain/catalog";
 import { DestinationSyncMode } from "core/request/AirbyteClient";
-import { BatchEditProvider, useBulkEdit } from "hooks/services/BulkEdit/BulkEditService";
-import { ConnectionFormMode, useConnectionFormService } from "hooks/services/ConnectionForm/ConnectionFormService";
-import { links } from "utils/links";
-import { naturalComparatorBy } from "utils/objects";
-import CatalogTree from "views/Connection/CatalogTree";
+import { useConnectionFormService } from "hooks/services/ConnectionForm/ConnectionFormService";
+import { CatalogTree } from "views/Connection/CatalogTree";
 
-import { BulkHeader } from "../../CatalogTree/components/BulkHeader";
-import Search from "./Search";
 import styles from "./SyncCatalogField.module.scss";
-
-const SubtitleCell = styled(Cell).attrs(() => ({ lighter: true }))`
-  font-size: 10px;
-  line-height: 12px;
-  border-top: 1px solid ${({ theme }) => theme.greyColor0};
-  padding-top: 5px;
-`;
-
-const ClearSubtitleCell = styled(SubtitleCell)`
-  border-top: none;
-`;
-
-const NextLineText = styled.div`
-  margin-top: 10px;
-`;
 
 interface SchemaViewProps extends FieldProps<SyncSchemaStream[]> {
   additionalControl?: React.ReactNode;
   destinationSupportedSyncModes: DestinationSyncMode[];
   isSubmitting: boolean;
-  mode?: ConnectionFormMode;
 }
 
-const CatalogHeader: React.FC = () => {
-  const { mode } = useConnectionFormService();
-  const { onCheckAll, selectedBatchNodeIds, allChecked } = useBulkEdit();
-  const catalogHeaderStyle = classnames({
-    [styles.catalogHeader]: mode !== "readonly",
-    [styles.readonlyCatalogHeader]: mode === "readonly",
-  });
-
-  return (
-    <Header className={catalogHeaderStyle}>
-      {mode !== "readonly" && (
-        <Cell className={styles.checkboxCell}>
-          <CheckBox
-            onChange={onCheckAll}
-            indeterminate={selectedBatchNodeIds.length > 0 && !allChecked}
-            checked={allChecked}
-          />
-        </Cell>
-      )}
-      {mode !== "readonly" && <Cell flex={0.2} />}
-      <Cell lighter flex={0.4}>
-        <FormattedMessage id="sources.sync" />
-      </Cell>
-      <Cell lighter>
-        <FormattedMessage id="sources.source" />
-        <InfoTooltip>
-          <FormattedMessage id="connectionForm.source.info" />
-        </InfoTooltip>
-      </Cell>
-      <Cell />
-      <Cell lighter flex={1.5}>
-        <FormattedMessage id="form.syncMode" />
-        <InfoTooltip>
-          <FormattedMessage id="connectionForm.syncType.info" />
-          <TooltipLearnMoreLink url={links.syncModeLink} />
-        </InfoTooltip>
-      </Cell>
-      <Cell lighter>
-        <FormattedMessage id="form.cursorField" />
-        <InfoTooltip>
-          <FormattedMessage id="connectionForm.cursor.info" />
-        </InfoTooltip>
-      </Cell>
-      <Cell lighter>
-        <FormattedMessage id="form.primaryKey" />
-        <InfoTooltip>
-          <FormattedMessage id="connectionForm.primaryKey.info" />
-        </InfoTooltip>
-      </Cell>
-      <Cell lighter>
-        <FormattedMessage id="connector.destination" />
-        <InfoTooltip>
-          <FormattedMessage id="connectionForm.destinationName.info" />
-          <NextLineText>
-            <FormattedMessage id="connectionForm.destinationStream.info" />
-          </NextLineText>
-        </InfoTooltip>
-      </Cell>
-      <Cell />
-    </Header>
-  );
-};
-
-const CatalogSubheader: React.FC = () => {
-  const { mode } = useConnectionFormService();
-
-  const catalogSubheaderStyle = classnames({
-    [styles.catalogSubheader]: mode !== "readonly",
-    [styles.readonlyCatalogSubheader]: mode === "readonly",
-  });
-
-  return (
-    <Header className={catalogSubheaderStyle}>
-      <SubtitleCell>
-        <FormattedMessage id="form.namespace" />
-      </SubtitleCell>
-      <SubtitleCell>
-        <FormattedMessage id="form.streamName" />
-      </SubtitleCell>
-      <SubtitleCell flex={1.5}>
-        <FormattedMessage id="form.sourceAndDestination" />
-      </SubtitleCell>
-      <ClearSubtitleCell />
-      <ClearSubtitleCell />
-      <SubtitleCell>
-        <FormattedMessage id="form.namespace" />
-      </SubtitleCell>
-      <SubtitleCell>
-        <FormattedMessage id="form.streamName" />
-      </SubtitleCell>
-    </Header>
-  );
-};
-
-const SyncCatalogField: React.FC<SchemaViewProps> = ({ additionalControl, field, form, isSubmitting }) => {
+const SyncCatalogFieldComponent: React.FC<React.PropsWithChildren<SchemaViewProps>> = ({
+  additionalControl,
+  field,
+  form,
+  isSubmitting,
+}) => {
   const { mode } = useConnectionFormService();
 
   const { value: streams, name: fieldName } = field;
 
-  const [searchString, setSearchString] = useState("");
   const setField = form.setFieldValue;
 
-  const onChangeSchema = useCallback(
+  const onStreamsUpdated = useCallback(
     (newValue: SyncSchemaStream[]) => {
       setField(fieldName, newValue);
     },
     [fieldName, setField]
   );
 
-  const onChangeStream = useCallback(
-    (newValue: SyncSchemaStream) => onChangeSchema(streams.map((str) => (str.id === newValue.id ? newValue : str))),
-    [streams, onChangeSchema]
-  );
-
-  const sortedSchema = useMemo(
-    () => streams.sort(naturalComparatorBy((syncStream) => syncStream.stream?.name ?? "")),
-    [streams]
-  );
-
-  const filteredStreams = useMemo(() => {
-    const filters: Array<(s: SyncSchemaStream) => boolean> = [
-      (_: SyncSchemaStream) => true,
-      searchString
-        ? (stream: SyncSchemaStream) => stream.stream?.name.toLowerCase().includes(searchString.toLowerCase())
-        : null,
-    ].filter(Boolean) as Array<(s: SyncSchemaStream) => boolean>;
-
-    return sortedSchema.filter((stream) => filters.every((f) => f(stream)));
-  }, [searchString, sortedSchema]);
-
   return (
-    <BatchEditProvider nodes={streams} update={onChangeSchema}>
-      <LoadingBackdrop loading={isSubmitting}>
-        <div className={styles.header}>
-          <Text as="h2" size="sm">
-            <FormattedMessage id={mode === "readonly" ? "form.dataSync.readonly" : "form.dataSync"} />
-          </Text>
-          {mode !== "readonly" && additionalControl}
-        </div>
-        {mode !== "readonly" && <Search onSearch={setSearchString} />}
-        <CatalogHeader />
-        <CatalogSubheader />
-        <BulkHeader />
-        <div className={styles.catalogTreeContainer}>
-          <CatalogTree streams={filteredStreams} onChangeStream={onChangeStream} />
-        </div>
-      </LoadingBackdrop>
-    </BatchEditProvider>
+    <>
+      <div className={styles.header}>
+        <Text as="h2" size="sm">
+          <FormattedMessage id={mode === "readonly" ? "form.dataSync.readonly" : "form.dataSync"} />
+        </Text>
+        {mode !== "readonly" && additionalControl}
+      </div>
+      <CatalogTree streams={streams} onStreamsChanged={onStreamsUpdated} isLoading={isSubmitting} />
+    </>
   );
 };
 
-export default React.memo(SyncCatalogField);
+export const SyncCatalogField = React.memo(SyncCatalogFieldComponent);
